pyproject.toml
[project]
name = "langextract-web-demo"
version = "0.1.0"
description = "Web demo for structured extraction using langextract + OpenAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "flask>=3.0.0",
  "langextract[openai]>=0.1.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"
